Upgrade to bigintegers in a few places

+++ datadrivendota/guilds/models.py
@@ -2,6 +2,6 @@ from django.db import models
 
 
 class Guild(models.Model):
-    steam_id = models.IntegerField(unique=True)
+    steam_id = models.BigIntegerField(unique=True)
     name = models.CharField(max_length=50)
     logo = models.BigIntegerField()
